feat(chat): unify guide+mingo chat behind useUnifiedChat with SSE/NATS adapters
