+++ openframe-frontend-core/src/components/chat/types/index.ts
@@ -5,3 +5,4 @@ export * from './network.types'
 export * from './processing.types'
 export * from './api.types'
 export * from './entities'
+export * from './unified-chat-state.types'
